python/requirements.txt
fastapi==0.109.0
pydantic==1.10.14
python-dotenv==1.0.1
uvicorn==0.30.1